# Improve handling of indefinite hessian matrices in fmintr