The trust region subproblems used in `fmintr` require the Hessian matrix to be positive definite. In the case that it isn't, we perform a modified Cholesky factorization such that the factors give a positive definite approximation to the true Hessian. The previous method for doing this modified factorization, while robust, is quite slow, likely because it's not taking advantage of optimized low level BLAS stuff.

This PR updates the modified Cholesky routine to use a bisection search to find a small diagonal correction to ensure the factors are positive definite. This is much faster in practice, often only 2-3x more expensive that a regular un-modified Cholesky factorization, and the results seem to be about as good on most problems I've tested with.

Some benchmarking for matrix of size 8000x8000:
```
Old cholmod
2min 3s ± 2.15 s per loop (mean ± std. dev. of 7 runs, 1 loop each)
New cholmod
6.25 s ± 119 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
Regular cholesky
1.83 s ± 67.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)